Updated comparing functionality

+++ src/Products/Comparison/Service/ComparisonServiceImpl.cs
@@ -17,9 +17,7 @@ using Newtonsoft.Json.Linq;
 namespace GroupDocs.Total.MVC.Products.Comparison.Service
 {
     public class ComparisonServiceImpl : IComparisonService
-    {
-        private static readonly string COMPARE_RESULT = "compareResult";
-        private static readonly string JPG = ".jpg";
+    {               
         private static readonly string DOCX = ".docx";
         private static readonly string DOC = ".doc";
         private static readonly string XLS = ".xls";
@@ -36,191 +34,6 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
         public ComparisonServiceImpl(GlobalConfiguration globalConfiguration)
         {
             this.globalConfiguration = globalConfiguration;
-        }
-
-        /// <summary>
-        /// Convert FormData object to CompareRequest object
-        /// </summary>
-        /// <param name="request">HttpRequest</param>
-        /// <returns></returns>
-        public CompareRequest GetFormData(HttpRequest request)
-        {
-            CompareRequest resultData = new CompareRequest();
-            resultData.files = new List<Stream>();
-            if (request.Files.Count > 2)
-            {
-                for (int i = 0; i < request.Files.Count; i++)
-                {
-                    using (var stream = new MemoryStream())
-                    {
-                        stream.Position = 0;//resetting stream's position to 0
-                        var serializer = new JsonSerializer();
-                        if (request.Files[i].FileName.Equals("blob"))
-                        {
-                            using (var sr = new StreamReader(request.Files[i].InputStream))
-                            {
-                                using (var jsonTextReader = new JsonTextReader(sr))
-                                {
-                                    JArray result = (JArray)serializer.Deserialize(jsonTextReader);
-                                    if (result.Count != 0)
-                                    {
-                                        List<CompareFileDataRequest> items = result.ToObject<List<CompareFileDataRequest>>();
-                                        if (items[0] != null)
-                                        {
-                                            if (items[0].GetFile().Contains("http") || items[0].GetFile().Contains("https"))
-                                            {
-                                                resultData.urls = items;
-                                            }
-                                            else
-                                            {
-                                                resultData.paths = items;
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        else
-                        {
-                            MemoryStream file = new MemoryStream();
-                            using (Stream currentStream = request.Files[i].InputStream)
-                            {
-                                currentStream.CopyTo(file);
-                            }
-                            file.Position = 0;                    
-                            resultData.files.Add(file);
-                        }
-                    }
-                }
-            }
-            else
-            {
-                for (int i = 0; i < request.Files.Count; i++)
-                {                   
-                    Stream stream = request.Files[i].InputStream;
-                    stream.Position = 0;                  
-                    resultData.files.Add(stream);
-                }
-                resultData.firstPath = (String.IsNullOrEmpty(request.Form["firstPath"])) ? request.Files["firstFile"].FileName : request.Form["firstPath"];
-                resultData.secondPath = (String.IsNullOrEmpty(request.Form["secondPath"])) ? request.Files["secondFile"].FileName : request.Form["secondPath"];
-                resultData.firstPassword = request.Form["firstPassword"];
-                resultData.secondPassword = request.Form["secondPassword"];
-            }
-            return resultData;
-        }
-
-        public string CalculateResultFileName(string documentGuid, string index, string ext)
-        {
-            // configure file name for results
-            string directory = globalConfiguration.Comparison.GetResultDirectory();
-            string resultDirectory = String.IsNullOrEmpty(directory) ? globalConfiguration.Comparison.GetFilesDirectory() : directory;
-            if (!Directory.Exists(resultDirectory))
-            {
-                Directory.CreateDirectory(resultDirectory);
-            }
-            string extension = (ext != null) ? GetRightExt(ext) : "";
-            // for images of pages specify index, for all result pages file specify "all" prefix
-            string idx = (String.IsNullOrEmpty(index)) ? "all" : index;
-            string suffix = idx + extension;
-            return string.Format("{0}{1}{2}-{3}-{4}", resultDirectory, Path.DirectorySeparatorChar, COMPARE_RESULT, documentGuid, suffix);
-        }
-
-        public bool CheckFiles(string firstFileName, string secondFileName)
-        {
-            string extension = Path.GetExtension(firstFileName);
-            // check if files extensions are the same and support format file
-            return extension.Equals(Path.GetExtension(secondFileName)) && CheckSupportedFiles(extension);
-        }
-
-        public bool CheckMultiFiles(List<string> fileNames)
-        {
-            string extension = Path.GetExtension(fileNames[0]);
-            // check if files extensions are the same and support format file
-            if (!CheckSupportedFiles(extension))
-            {
-                return false;
-            }
-            foreach (string path in fileNames)
-            {
-                if (!extension.Equals(Path.GetExtension(path)))
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        public CompareResultResponse Compare(CompareRequest compareRequest)
-        {
-            string firstPath = compareRequest.firstPath;
-
-            ICompareResult compareResult;
-            // create new comparer
-            Comparer comparer = new Comparer();
-            // create setting for comparing
-            ComparisonSettings settings = new ComparisonSettings();
-
-            // compare two documents
-            compareResult = comparer.Compare(firstPath,
-                compareRequest.firstPassword,
-                compareRequest.secondPath,
-                compareRequest.secondPassword,
-                settings);
-
-            if (compareResult == null)
-            {
-                throw new Exception("Something went wrong. We've got null result.");
-            }
-            string saveTemp = null;
-            if (Path.GetExtension(firstPath).Equals(".html") || Path.GetExtension(firstPath).Equals(".htm"))
-            {
-                saveTemp = Path.Combine(globalConfiguration.Comparison.GetResultDirectory(), "temp.html");
-            }
-            // convert results
-            CompareResultResponse compareResultResponse = GetCompareResultResponse(compareResult, saveTemp);
-
-            //save all results in file
-            string extension = Path.GetExtension(firstPath);
-            SaveFile(compareResultResponse.GetGuid(), null, compareResult.GetStream(), extension);
-            File.Delete(saveTemp);
-            compareResultResponse.SetExtension(extension);
-
-            return compareResultResponse;
-        }
-
-        public CompareResultResponse CompareFiles(Stream firstContent, string firstPassword, Stream secondContent, string secondPassword, string fileExt)
-        {
-            ICompareResult compareResult;
-            // create new comparer
-            Comparer comparer = new Comparer();
-            // create setting for comparing
-            ComparisonSettings settings = new ComparisonSettings();
-
-            // compare two documents
-            compareResult = comparer.Compare(firstContent,
-                    firstPassword,
-                    secondContent,
-                    secondPassword,
-                    settings);
-
-            if (compareResult == null)
-            {
-                throw new Exception("Something went wrong. We've got null result.");
-            }
-            string saveTemp = null;
-            if (fileExt.Equals(".html") || fileExt.Equals(".htm"))
-            {               
-                saveTemp = Path.Combine(globalConfiguration.Comparison.GetResultDirectory(), "temp.html");
-            }
-            // convert results
-            CompareResultResponse compareResultResponse = GetCompareResultResponse(compareResult, saveTemp);
-
-            //save all results in file
-            SaveFile(compareResultResponse.GetGuid(), null, compareResult.GetStream(), fileExt);
-            File.Delete(saveTemp);
-            compareResultResponse.SetExtension(fileExt);
-
-            return compareResultResponse;
         }
 
         public List<FileDescriptionEntity> LoadFiles(PostedDataEntity fileTreeRequest)
@@ -251,7 +64,7 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
                 {
                     FileInfo fileInfo = new FileInfo(file);
                     // check if current file/folder is hidden
-                    if (fileInfo.Attributes.HasFlag(FileAttributes.Hidden) || 
+                    if (fileInfo.Attributes.HasFlag(FileAttributes.Hidden) ||
                         Path.GetFileName(file).Equals(Path.GetFileName(globalConfiguration.Comparison.GetFilesDirectory())))
                     {
                         // ignore current file and skip to next one
@@ -281,14 +94,90 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
             }
         }
 
-        public LoadDocumentEntity LoadDocumentPages(PostedDataEntity loadResultPageRequest)
+        //public string CalculateResultFileName(string documentGuid, string index, string ext)
+        //{
+        //    // configure file name for results
+        //    string directory = globalConfiguration.Comparison.GetResultDirectory();
+        //    string resultDirectory = String.IsNullOrEmpty(directory) ? globalConfiguration.Comparison.GetFilesDirectory() : directory;
+        //    if (!Directory.Exists(resultDirectory))
+        //    {
+        //        Directory.CreateDirectory(resultDirectory);
+        //    }
+        //    string extension = (ext != null) ? GetRightExt(ext) : "";
+        //    // for images of pages specify index, for all result pages file specify "all" prefix
+        //    string idx = (String.IsNullOrEmpty(index)) ? "all" : index;
+        //    string suffix = idx + extension;
+        //    return string.Format("{0}{1}{2}-{3}-{4}", resultDirectory, Path.DirectorySeparatorChar, COMPARE_RESULT, documentGuid, suffix);
+        //}       
+
+        public bool CheckFiles(CompareRequest files)
+        {
+            string extension = Path.GetExtension(files.guids[0].GetGuid());
+            // check if files extensions are the same and support format file
+            if (!CheckSupportedFiles(extension))
+            {
+                return false;
+            }
+            foreach (CompareFileDataRequest path in files.guids)
+            {
+                if (!extension.Equals(Path.GetExtension(path.GetGuid())))
+                {
+                    return false;
+                }
+            }
+            return true;
+        }
+
+        public CompareResultResponse Compare(CompareRequest compareRequest)
+        {
+            string firstPath = compareRequest.guids[0].GetGuid();
+
+            ICompareResult compareResult;
+            // create new comparer
+            Comparer comparer = new Comparer();
+            // create setting for comparing
+            ComparisonSettings settings = new ComparisonSettings();
+            settings.StyleChangeDetection = true;
+            // compare two documents
+            compareResult = comparer.Compare(firstPath,
+                compareRequest.guids[0].GetPassword(),
+                compareRequest.guids[1].GetGuid(),
+                compareRequest.guids[1].GetPassword(),
+                settings);
+
+            if (compareResult == null)
+            {
+                throw new Exception("Something went wrong. We've got null result.");
+            }
+            //string saveTemp = null;
+            //if (Path.GetExtension(firstPath).Equals(".html") || Path.GetExtension(firstPath).Equals(".htm"))
+            //{
+            //    saveTemp = Path.Combine(globalConfiguration.Comparison.GetResultDirectory(), "temp.html");
+            //}
+            // convert results
+            //save all results in file
+            string extension = Path.GetExtension(firstPath);
+            CompareResultResponse compareResultResponse = GetCompareResultResponse(compareResult, extension);
+
+           
+           // SaveFile(compareResultResponse.GetGuid(), null, compareResult.GetStream(), extension);
+            //if (!String.IsNullOrEmpty(saveTemp))
+            //{
+            //    File.Delete(saveTemp);
+            //}
+            compareResultResponse.SetExtension(extension);
+
+            return compareResultResponse;
+        }
+               
+        public LoadDocumentEntity LoadDocumentPages(string path)
         {
             LoadDocumentEntity loadDocumentEntity = new LoadDocumentEntity();           
             //load file with results
             try
             {
                 Comparer comparer = new Comparer();
-                List<PageImage> resultImages = comparer.ConvertToImages(loadResultPageRequest.path);
+                List<PageImage> resultImages = comparer.ConvertToImages(path);
 
                 foreach (PageImage page in resultImages) {
                     PageDescriptionEntity loadedPage = new PageDescriptionEntity();
@@ -344,15 +233,17 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
             // convert results
             CompareResultResponse compareResultResponse = GetCompareResultResponse(compareResult, saveTemp);
 
-            //save all results in file
-            SaveFile(compareResultResponse.GetGuid(), null, compareResult.GetStream(), ext);
-            File.Delete(saveTemp);
+           
+            if (!String.IsNullOrEmpty(saveTemp))
+            {
+                File.Delete(saveTemp);
+            }
             compareResultResponse.SetExtension(ext);
 
             return compareResultResponse;
         }
 
-        private CompareResultResponse GetCompareResultResponse(ICompareResult compareResult, string saveTemp)
+        private CompareResultResponse GetCompareResultResponse(ICompareResult compareResult, string ext)
         {
             CompareResultResponse compareResultResponse = new CompareResultResponse();
 
@@ -362,51 +253,35 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
 
             string guid = System.Guid.NewGuid().ToString();
             compareResultResponse.SetGuid(guid);
-
-            // if there are changes save images of all pages
-            // unless save only the last page with summary
-            Stream[] images = null;
-            if (changes != null && changes.Length > 0)
+            //save all results in file
+            string resultGuid = SaveFile(compareResultResponse.GetGuid(), compareResult.GetStream(), ext);
+            List<PageDescriptionEntity> pages = LoadDocumentPages(resultGuid).GetPages();
+            List<string> pageImages = new List<string>();
+            foreach (PageDescriptionEntity page in pages)
             {
-                if (!String.IsNullOrEmpty(saveTemp))
-                {                                     
-                    compareResult.SaveDocument(saveTemp);
-                    images = compareResult.GetImages();                    
-                }
-                else
-                {
-                    images = compareResult.GetImages();                   
-                }
-                List<string> pages = SaveImages(images, guid);
-                // save all pages
-                compareResultResponse.SetPages(pages);
+                pageImages.Add(page.GetData());
             }
-            else
-            {
-               images = compareResult.GetImages();
-                int last = images.Length - 1;
-                // save only summary page
-                compareResultResponse.AddPage(SaveFile(guid, last.ToString(), images[last], JPG));
-            }
+            compareResultResponse.SetPages(pageImages);
+            compareResultResponse.SetGuid(resultGuid);
             return compareResultResponse;
         }
 
-        private List<string> SaveImages(Stream[] images, string guid)
-        {
-            List<string> paths = new List<string>(images.Length);
-            for (int i = 0; i < images.Length; i++)
-            {
-                paths.Add(SaveFile(guid, i.ToString(), images[i], JPG));
-            }
-            return paths;
-        }
+        //private List<string> SaveImages(Stream[] images, string guid)
+        //{
+        //    List<string> paths = new List<string>(images.Length);
+        //    for (int i = 0; i < images.Length; i++)
+        //    {
+        //        paths.Add(SaveFile(guid, i.ToString(), images[i], JPG));
+        //    }
+        //    return paths;
+        //}
 
-        private string SaveFile(string guid, string pageNumber, Stream inputStream, string ext)
+        private string SaveFile(string guid, Stream inputStream, string ext)
         {
-            string imageFileName = CalculateResultFileName(guid, pageNumber, ext);
+            string fileName = Path.Combine(globalConfiguration.Comparison.GetResultDirectory(), guid + ext);
             try
             {
-                using (var fileStream = File.Create(imageFileName))
+                using (var fileStream = File.Create(fileName))
                 {
                     inputStream.Seek(0, SeekOrigin.Begin);
                     inputStream.CopyTo(fileStream);
@@ -417,7 +292,7 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
             {
                 throw new Exception("Exception occurred while write result images files.");
             }
-            return imageFileName;
+            return fileName;
         }
 
         /// <summary>
@@ -448,7 +323,7 @@ namespace GroupDocs.Total.MVC.Products.Comparison.Service
                 case ".html":
                     return HTML;
                 case ".htm":
-                    return HTM;
+                    return HTM;                    
                 default:
                     return ext;
             }
